+++ packages/widgetbook/lib/src/addons/builder_addon/addon.dart
@@ -1,0 +1,1 @@
+export 'builder_addon.dart';