+++ sandbox/lib/widgetbook.dart
@@ -50,9 +50,8 @@ class WidgetbookApp extends StatelessWidget {
             DefaultMaterialLocalizations.delegate,
           ],
         ),
-        UseCaseBuilderAddon(
+        BuilderAddon(
           name: 'Bounds',
-          isEnabled: false,
           builder: (context, child) => Container(
             decoration: BoxDecoration(
               border: Border.all(